Splitting
- Work on splliting renderer into parser and renderer classes
