+++ example/index.php
@@ -2,11 +2,19 @@
 
 require_once '../src/DBlackborough/Quill/Renderer.php';
 require_once '../src/DBlackborough/Quill/Renderer/Html.php';
+require_once '../src/DBlackborough/Quill/Parser.php';
+require_once '../src/DBlackborough/Quill/Parser/Html.php';
 
 $deltas = '{"ops":[{"insert":"Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed efficitur nibh tempor augue lobortis, nec eleifend velit venenatis. Nullam fringilla dui eget lectus mattis tincidunt. \nDonec sollicitudin, lacus sed luctus ultricies, "},{"attributes":{"strike":true,"italic":true},"insert":"quam sapien "},{"attributes":{"strike":true},"insert":"sollicitudin"},{"insert":" quam, nec auctor eros felis elementum quam. Fusce vel mollis enim. \n\n"},{"attributes":{"bold":true},"insert":"Sed ac augue tincidunt,"},{"insert":" cursus urna a, tempus ipsum. Donec pretium fermentum erat a "},{"attributes":{"underline":true},"insert":"elementum"},{"insert":". In est odio, mattis sed dignissim sed, porta ac nisl. Nunc et tellus imperdiet turpis placerat tristique nec quis justo. Aenean nisi libero, auctor a laoreet sed, fermentum vel massa.\n\nEtiam ultricies leo eget purus tempor dapibus. Integer ac sapien eros. Suspendisse convallis ex."}]}';
 
-$renderer = new \DBlackborough\Quill\Renderer\Html();
 
-if ($renderer->load($deltas) === true) {
-    var_dump($renderer->render());
-}
+$parser = new \DBlackborough\Quill\Parser\Html();
+$parser->load($deltas);
+
+var_dump($parser->load($deltas));
+
+var_dump($parser->parse());
+
+$renderer = new \DBlackborough\Quill\Renderer\Html($parser->content());
+
+print $renderer->render();
